Add Bloc providers for layout and profile management
Introduce `LayoutCubit` and `ProfileCubit` to manage the state of the layout and profile features respectively. The `ProfileCubit` is initialized with `ProfileRepoImp` and fetches the profile data upon creation. These providers are wrapped in a `MultiBlocProvider` to ensure they are available throughout the app. This change enhances the app's state management capabilities and prepares for future feature integrations.